"All the details Rudder has are a definite red flag. I cannot really conclude for sure that the reward decomposition is responsible for the superior performance. I think a real baseline would be to take the LSTM they used and instead of decomposing the rewards to individual actions they used it as a critic directly. Thinking about it logically, for the LSTM to redistribute the reward to an earlier action first it has to know that the action caused it. And if it caused it, it should also show up in the value function."
I am going to ask the author about this tomorrow. With the information in the paper I cannot conclude for sure that reward decomposition has anything to do with the superior performance of Rudder on the two tasks.

For now I am done for the day. I should be done with my studies tomorrow.

I am actually pretty sure that my reasoning with regard to Rudder is right, though I want to make a full pass through the paper before I fire a message. And if it really turns out to be right then that will vastly simplify things for me.