"Hello. I've finally went through enough of my workload that I have both room in my head and the time to make a fuller analysis of Rudder in anticipation of implementing it. I've looked at the paper previously, but not like now where I dove into it to the point of thinking through all the details that would be needed for an actual implementation.
In the process of doing that I've realized that the paper has a very important omission from the perspective of an aspiring practitioner such as myself. Let me just say that I believe everything the paper claims and that includes the deep learning results on the games Venture and Bowling.

What the paper does not claim however is that those results are solely due to the reward redistribution. When I studied the paper previously I inferred that they were from the context, but now that I am in business mode I see that there is an entirely different interpretation of those results. As unlikely as it seems, there is nothing to indicate that you haven't exceeded the state of the art by as much as 20x simply because the LSTM you were using to do reward redistribution is a very good critic.

Saying this might be strange as AC agents are known to be less efficient than DQNs, but it is not beyond reason to think that if the critic they are using is much better that they could exceed the DQNs in sample efficiency even by as much as 20x. If this interpretation is true then this might be an important finding in itself.

Logically speaking, the LSTM cannot possibly do the reward redistribution to some earlier action until it learns enough to do so. And once it knows to do so, that information should manifest in the cost function it is optimizing. If that is wrong that would mean that the contribution analysis has more precise insight than the cost function. I do not see this as impossible, but it is highly unlikely. Hence it should be possible to get the same performance on Venture and Bowling by using the LSTM as the critic directly.

Have any of you tested this?

Because the contribution analysis would add such a great deal of implementation and computational complexity, I would like to see such an ablation study. In lieu of a study, I'd be happy to know anything useful on an informal basis. Thank you."

I sent this to one of the authors and I'll see how it goes from here.

Ugh, today feel like it will be fucked up. I'll have an appointment later in the day. Today I've been in action since 6:50am and only just returned half an hour ago.

To be honet, I think that my suspicions are right here - if he gets back to me saying that they in fact have the results that redistribution works better than the regular thing or with some piece of reasoning showing that it should work better I will be rather surprise.

Hence, I think I've saved myself quite a lot of work by doing this.

The Rudder's second interpretation is a lesson in just how valuable critics really are. The lessons of the last few days is something I will be carrying well into my career.

I did not do a lot of hard programming since the 21st, but I sure did a lot of hard thinking.

With this everything is set - I'll temporarily and perhaps permanently discharge my obligation to implement Rudder and finish the work on the modular plastic RNN.

After that everything will be set for me. I mocked the OpenAI guys, but they are not wrong in terms of direction. They are just off target.

In less than 3 weeks I will have everything I need to create the elite agents.

I'll also opportunistically start looking an data science contests. I'll finally have an edge with the new architecture on something and will be able to cash in on it somewhere. It is not like I need to do borderline illegal things as my vocation.

At any rate, I won't be doing programming today probably. Let me see if I have anything else to do first.