I am finally done with those annoying appointments. Phew. I do not feel like programming right now though. I am still thinking what comes next.
I really wish I'd get a reply from /u/gerdi sooner as it would lift a load from my mind. It is exhausting to keep dual hyptheses in it.

I've been thinking what should be my main goal from here, and my thoughts are going in the direction of making it harder for static parts to do any memorization. I should apply the state difference trick to every layer in the network apart from the two heads.

Backprop should only be used to establish the control system. All the memorization should be done through unsupervised Hebbian learning.

This will allow long term memories longer than the number of steps allowed to via backprop.

The thing is, backprop can only memorize as far as the number of steps allowed for it. With 250 steps at most it can emulate a circuit unrolled for such a number of steps. With only static parts that would be its limit.

To induce more, it will be necessary to put in priors.

A lot would be possible with a network of a few hundred layers unrolled in time. I actually have a higher regard for backprop than what the brain does. Backprop is unlimited, but the brain can only work in the tiny box it is confined in by nature.

Urgh...damn I do not feel like programming. Programming is something I need to get into the mood first. These trips outside the house break my rythm.

Since the fall season is starting and weather has turned cold, it is time to start preparing for winter. Let me do those chores instead. I need to get some exercise.