I am still thinking about it. If I really just want those views with a minimum of fuss, then I should just write a wrapper.
Since I am in designy mindset right now, what I am going to is redesign the whole library starting from tomorrow.

First, I will make the `HostTensor` a base with scalar dimensions - no more `{from near_to}`. Yeah, I know that I had an argument with myself about this, but it turns out we were both wrong. I should just have made the standard tensor the most basic and then write a wrapper around it.

Even if I used the ranges a total of one time through the entire span of dragging it, I still think a serious tensor type should support this. But it true that having to deal with range dimensions is a constant drag on everything.

And now I do not have an easy way to just extend this. It is ironic - now that I have an actual use for tensor views they are completely useless.

...Maybe it is time to go through another round of redesign. This will be a pain, but this is something I've been long used to. I need to do it now, later Spiral might have actual users and decision like this will be impossible. I need to optimize for commonly used cases.

Once I do this, that matrix initializer will be much easier to add. I'll split the current `HostTensor` into `HostTensor` and `ViewHostTensor`.

Now I can't decide whether I should simply the kernels as well. Do I want to allow only basic tensors, or a subset of view tensors like I do now? Agh...

...Restrict it. Restrict it unless it is necessary. Simplify things not just for the users, but in the library as well. Throwing away work is fine, especially work that might be useful, but wasn't for months.

I'll start work on this tomorrow. It will most likely take me a few days to get through a major redesign such as this, but it should not be as bad as the thing with the Cuda kernels which took me over a week.

The fact that I need the named views in order to support safe fusion is absolute. After I do this, whole of Spiral will be another step closer towards being complete, not just as a language, but as the ecosystem supporting my goals and the foundation of my power.