My fatigue is gone, but I do not feel like starting work immediatelly. Yeah, these redesigns are a pain in the ass, but I should go through it.
Well, let me read GS and a chapter of than LN I am reading and then I will start.

I need to get this done and then I will be able to move to the next step. I am still wondering when I will get a reply to my question on Rudder.

I've noticed from my interaction with academics that they are like politicians. They will be happy to take in praise and field the easy questions, but anything hard or what would require effort to answer they will just take advantage of their right to ignore.

But I guess 2 out of 4 with PRONG and DP would not be bad. And with Rudder, I can always just take the state difference trick from it - the static poker player is just too easy of an target, but I am guessing in some other situations it would be quite good. I will have a better chance to test when I do the Maze test.

Edit: Forgot to push this. Let me start.

Actually, I need to focus on this for a bit more. I just realized that the inner part of the tensor is what I want to use named indexing for.

Let me just sketch things out as much as possible first. It is not a given that I am going redesign everything.