"4:05pm. I am still thinking about it.
```
        apply = inl data i ->
            Tuple.foldl (inl {data with basic dim} ->
                function
                | .(_) ->
                    match dim with
                    | () -> error_type "Cannot apply the tensor anymore."
                    | {$k=dim} -> {data with dim basic = self .view dim}
                | i ->
                    match dim with
                    | () -> error_type "Cannot apply the tensor anymore."
                    | from :: dim -> {data with basic = self (i+from); dim}
                ) data (Tuple.wrap i)
            |> facade
```

I am considering putting the machinery for `view` into apply for the `HostTensor`.

When passed in standard ints, it would act as an application. But passing in `()` would act as an empty view, and `{from (near_to | by)}` would also act as views.

And I could remove what `view` was before. I do need those in general. Creating the `ViewTensor` should suffice for that.

That `()` is actually a great idea.

Originally I intended to index into named views using the `.input` or `.ouput` kind of syntax, but now it occurs to me that I would be better off using the tree based indexing like I've used in other parts of the ML library.

5:05pm. I am still thinking about it. Today seems to have turned out into a brainstorming session. I am just sitting here on this comfy chair while listening to music and wracking my brain over seemingly trivial design issues. To be honest, I am quite irritated right now. I want to bring the whole Rudder thing to a close rather than letting it fester in my mind. I am going to give it two more days and then I will just email the first on the author list to ask him about this.

If I do not get a reply within two days after that I'll just assume that my theory is right and wash my hands of this.

Analysis work is quite stressful.

5:10pm. I think I quite like what I have in mind for the new tensor.

The view tensor will have two kind of views. The range view will allow any upper and lower bound, while the tree view I will allow only the size. That will simplify things to the utmost.

5:20pm. How about I call it a day here? Obviously, I will not be programming anymore today and I think I've covered everything I wanted for the time being. I think now that I actually have a workable plan of action rather than just a desire to do it, I should have an easier time putting my ideas into code tomorrow."