https://www.kaggle.com/c/two-sigma-financial-news
This might be an interesting use for the RNN, but honestly, the prizes here are so damn cheap given the nature of the task. It is a waste of my time.

https://www.kaggle.com/c/quickdraw-doodle-recognition

This one...honestly, this one feels like it would be up my alley. Unfortunately 2 months is just too early for me to get it all online. I do not have the infrastructure to do metalearning for images and have no interest to make it at the moment.

I've considered Kaggle, but looking at the kinds of things that are here, I'd be better off solving the tasks that I want for my own sake.

It is annoying. The number of things I can dream about to make large gains in RL seems to be steadily decreasing to be replaced by the great unknown that is the metalearning.

Looking over the past 3 months I can only feel a quiet simmering inside me. 'Are these people trying to take me for a fool?' I have that kind feeling inside of me right now.

And though I am annoyed at him, I can only be grateful to Miconi for linking me to that reversible learning paper.

And even though Rudder is misleading, all the reasoning that can be applied to reward redistribution can also be applied to critics. Those figures are not a lie - a good critic can get that much.

The question is not whether the actor or the critic, but how to get the flow state going where the critic directs the actor by acting on its intentions.

It is actually quite rate to have a clear cut case of where rewards should be distributed. It is much easier to reward and punish based on the flow states.

Unsupervised learning is the key to that, and the simplest incarnation of it is the one that I have now.

Once I am done with the stupid tensor redesign, the great project will start - the hunt for the Inspired Desire.

---

The paper I read a few days ago, the one which use MAML to do model based learning did inspire me a little.

I am thinking that to attain true success I need to make model based learning intrinsic to the model, rather than a goal.

The metalearning that the plastic RNN can do is a step up from MAML, but to attain true victory I will need to take a step after that.

Maybe it will turn out to be the case that the plastic RNN + the state difference trick at every layer will do it, but I doubt it will be that easy. They will be necessary, but not sufficient.

I am done with the training phase for real. I will forget about Rudder and Zap, take the cards in my hand and play them for all they are worth. I've earned this for all the work I've put in since April. Rather than dwell on past victories and failures it is time to start focus on the next steps. Once I do that I will start looking forward to it and the motivation to move will come to me.

In poker as in life, aggression will not win a battle on its own, but aggression is a part of every winning strategy.

Never once have I lost sight of that.