"8:15am. I've been thinking about Rudder for the last two hours. I was wrong. When the rewards are stochastic then even the optimal critic will have a lot of variance with delayed rewards. Damn.
Let me send a second message to one of the authors.

9:05am.

```
It took me 3 days, but I finally figured out that even with an optimal critic, if the rewards are stochastic then the variance will be large when training an AC with a MC critic. I was not thinking about the rewards being stochastic when I wrote the earlier message - only transitions which could be modeled after. I was wrong - it seems my initial impression of Rudder was correct after all.

Unless a radically different way of doing credit assignment comes along, I will have no choice but to implement it at some point unfortunately.

The reason why this is unfortunate is because the current contribution analysis methods are highly unwieldy.

It is even worse than when the paper first came out - with the kinds of plastic RNNs that I want to use, each pass of integrated gradients would cost as much as a regular one and parallelism would not help. So having 32 integration steps would be like having to go through 32 layers for the critic. Ouch.

Layerwise relevance propagation would require additional implementation and is brittle according to various papers that I've read.

What finite difference approximation is to sensitivity analysis, integrated gradients are to contribution analysis. I'd very much like a backprop equivalent to it instead.

Well, that much is my problem. Maybe somebody will come up with a better contribution analysis method at some point. Let me move to the next questions:

1) In the paper you've trained the LSTM with the goal of predicting the total reward plus some auxiliary tasks. This is reminiscent of MC training of a regular critic. I've been wondering if instead of doing that, using dilated connections and training the LSTM to predict individual rewards would be better?

Since you are not using the value function for the critic directly and instead do a contribution analysis there is no need to train it as a critic. Maybe a supervised approach would have lower variance?

2) On page 8 you say: "To make static objects visible, we augment the input with the current frame." Doesn't that undo the effect of taking the difference of states, at least for the Atari games where the states are pixelwise inputs?

3)  The LSTM prediction g is decomposed by integrated gradients: "g(x) = sum {t=0; T} (h(at, ∆(st, st+1)) + 1/(T + 1)g(0))."

I am not sure what this is. Is `g(x)` supposed to be the total return prediction? If so what is `1/(T + 1)g(0))` supposed to be doing? Also `sum {t=0; T} (h(at, ∆(st, st+1))` probably means that you are making a prediction at each timestep. I thought that the total return was predicted only at the end of an episode?

4) I think you could go further in your theory. I do not see why it is necessary to convert the MDP to a delayed reward MDP and then redistribute from that. Is there any reason why redistributing rewards for any finite immediate reward MDP would not work?

5) Despite what I said originally, I still think an ablation study would be good. Even though I can now see how what Rudder does is different from an optimal critic, it is still difficult for me to imagine how things would behave with an optimal critic. Yes, the delayed (stochastic) rewards would cause variance that would be unavoidable without Rudder, but the optimal critic would also correct for them over time. When combined with natural gradient methods which I intend to do would that end up matching Rudder? How would fight of optimal critic vs optimal reward redistribution look like?

In your paper you do a lot more than train a bog standard critic. In the first place, an LSTM critic is rather rare - the Rudder paper is the first time I've actually seen an RNN used specifically for that and you do safe exploration and a prioritized replay buffer which are also innovations in their own right. Those DQNs are not the right baseline for it.

6) "Occasional strong artifacts due to numerical instabilities via strong non-linearities were observed at the last time steps of long episodes. These artifacts were removed by setting the integrated gradients signal for the last 10 timesteps of an
episode to 0."

Is this due to the tanh's and the sigmoids in the LSTM? What if you replaced them with their piecewise linear equivalents? Tanh could also be replaces with Relu+LN.

Let me just say that while these hacks might be fine for getting a paper out the door, they really make me antsy about using Rudder as a part of a real system - because as a rule I cannot afford to spend time tailoring what should be done internally by the contribution analysis to the task at hand. The necessity of what is described in these and the following sentences are the ones which cemented my view that a better contribution analysis method would be needed for Rudder to shine.

7) "Furthermore, the quality of the integrated gradients and the prediction error of the LSTM were used
to estimate the reliability of the reward redistribution. Reward redistribution is only used if both the
prediction and the backward analysis were judged to be reliable for the current episode..."

How bad is the reward redistribution when it is really bad? What do you use when when no redistribution is done? I am guessing either MC or no training is done at all.

---

Thank you for your time.
```

Let me send this. I'll see what I will get. I admit I am still trying to find a way to squirm out of my obligation. Rudder is definitely in the right direction, but I need to do even better than it.

9:55am. https://openreview.net/forum?id=r1lyTjAqYX
Recurrent Experience Replay in Distributed Reinforcement Learning

Let me read this and then I will start programming. That new tensor won't take care of itself. I am the one that is going to have to do it."