https://openreview.net/forum?id=S1en0sRqKm On the Computational Inefficiency of Large Batch Sizes for Stochastic Gradient Descent
This paper seems interesting.

https://openreview.net/forum?id=H1gsz30cKX
The Unreasonable Effectiveness of (Zero) Initialization in Deep Residual Learning

This one too.

I sent a mail to Martha White asking her for references on her talk.

Honestly, I have no idea what I am going to do about Rudder. The more I think about it, the more I like the theory of it and the less I like the whole contribution analysis thing tied to it.

I've been reading some stuff on deep Taylor Decomposition which is related to LRP. I mean as a method to maybe explain the visual inputs it might hold some muster, but there is no way I feel like dealing with its arbitrary rules - it requires relus and positive weights and what not. It would also add a very non-trivial amount of complexity to the library. Integrated gradients on the other hand would break the bank on the computation budget and would not be as simple as they seem.

Backprop is not something I can run willy nilly - it is mutable for a reason. I remember when I tried Reptile how much a few extra iteration hurt the runtime.

And now that R2D2 paper came out show Rudder tier gains just by using the LSTM for the Q network. This is exactly what I asked the author 3 days ago and it is like the sequence of events is trying to push me in a certain direction. But coincidences do not change the logic of anything - my reasoning for both liking and hating Rudder does not fundamentally change.

UAAhhhhhhhhhhhhhhhhhhhhh...

I have no idea. I have really no idea.

But one idea which keeps recurring would be - train the critic using MC and train the actor using TD. That could be one way to fight both variance and bias. It is not quite reward decomposition, but it might turn out to be close enough in practice. It would be easy to implement and would save me from hacks.

I need to think about this. Maybe I'll ask the author of that TD or not TD paper about it.

Let me go get some exercise now. I've been in front of the screen for long enough.