Backpropamine: training self-modifying neural networks with differentiable neuromodulated plasticity https://openreview.net/forum?id=r1lrAiA5Ym
Oh, this seems like it will be a interesting paper. Ah, here it is. This is probably by the Uber group. Surprisingly it is on LSTMs.

https://developer.nvidia.com/cuda-toolkit/whatsnew

Cuda 10 is also out.

https://arxiv.org/pdf/1705.10257.pdf
Boltzmann Exploration Done Right

This might be interesting too.

Basically, there is no way I am going to be doing any programming today anymore. What I want to do from here is do more research. I am starting to converge on the idea of training the actor using TD and the critic using MC. That should combine the strengths of the two approaches. I am sort of using the Rudder paper to argue against it now.

I want to review that TD or not TD paper again. Maybe it covered this issue. I did not get what it is doing in its entirety.

I think I will spend tomorrow reading the latest papers and finally resolving this. I've been under a severe block in the last few days.

Once I remove it, I will get push through this in one blast. The first thing I will do when I am done with the tensor rewrite will be testing out the tdmc actor critic.