I am ready to start the day. Right now I am thinking where I should start. Probably by posting the backpropamine paper on the ML sub.
After I am done with that, I guess I'll start reading those papers. I am starting to get my focus back after getting Rudder'd. I really think the TD for the actor and MC for the critic might be the right answer.

Regardless of the case, relevance propagation is absolutely out of the question. And I really want something better than IG.

I've been thinking of how an optimal critic vs an optimal redistributor would play out.

An optimal MC critic would be worse than a redistributor, but and optimal TD critic would be better. That is the cold hard truth. I only figured out that an optimal TD critic would be better last night.

Insights come to me overnight and in tiny fragments it seems.

But understanding this is important as it resolves a dilemma that was tearing me apart for the last few days. I'll just use an RNN critic (maybe the way I am doing it now would suffice too) and then use TD to update the actor.