https://www.reddit.com/r/MachineLearning/comments/9jukmz/rbackpropamine_training_selfmodifying_neural/
"Paper link: https://openreview.net/forum?id=r1lrAiA5Ym

I honestly think that the research direction the original [differentiable plasticity paper](https://arxiv.org/abs/1804.02464) opened is more revolutionary than it seems at first glance.

Not only does it have great tie ins to biologically plausible learning, it is the first time I saw a RNN being used to update another network's weights. Because now a successful example of that exists, a vast range of research into biologically plausible learning rules and not just Hebbian learning is suddenly live and could potentially be used to great effect as a part of deep learning.

Right now, all metalearning approaches have significant flaws:

1) MAML - In addition to being expensive, it has the large disadvantage of introducing the artificial metatraining/training separation.

2) SNAIL - It is a particularly large architecture.

3) Plasticity - It requires copying the plastic weights at every timestep in order to backprop through it. This is a huge memory expenditure.

In case of plasticity it is actually possible to work around its disadvantages by reversing the weight updates. I've done a little numerical stability testing of how various update rules behave in reverse mode and have discovered that while the EMA and Oja's rules from the original paper are stable for 1k-10k steps after which the weights explode due to exponential accumulation of error, weight norm can be reversed for 100s of thousands of timesteps with only tiny error even when float32 types are used.

When [reversible computing](https://www.youtube.com/watch?v=IQZ_bQbxSXk) finally arrives it will have significant implications for learning.

There was [a reversible learning paper](https://arxiv.org/abs/1502.03492) from 2015 which used it to optimize hyperparameters that I now see in an entirely different light. It is really too bad this was not tried for networks in networks.

That having said simple local unsupervised learning rules like for Hebbian learning have their advantages in that they can be easily reversed and would not require higher order differentiation for that to be done.

The Backpropamine paper is actually quite surprising to me in its content - in my own testing on Binary Pattern I've tried modulating the weight updates directly as was done in the paper, but I've found it to be highly unstable and eventually removed it. That might have been because I was modulating the plasticity coefficients in tandem with that so it might be worth revisiting. At least on the Binary Pattern, I've found modulating the plasticity coefficients to work very well - the plastic RNN from the original paper takes over 1k iterations to get to 99% sign accuracy on the Binary Pattern test. The architecture that modulates the plasticity coefficients managed to get to that point in about 60.

But that is just one sample on a toy task - there is a lot of work to be done before this line of research comes to fruition and it should be interesting follow as plasticity enables capabilities that would otherwise be impossible in regular RNNs. I'd like to see this tried on Dota and Starcraft. A recent [R2D2 paper](https://openreview.net/forum?id=r1lyTjAqYX) came out showing great gains in using a LSTM for learning value-action functions and better memories in a RNN will only extend that advantage."