What a waste of time - there is nothing I really want to post apart from those two papers that I already did. Also, I can't believe I am even going to have to test the TD idea. Rudder is supposed to be benefit, but it took a lot out of me. I guess I'll take my time slowly unravelling the mystery it dropped into my lap.
There is really only so much I can figure out just by thinking about it. My brain was really in high gear for the last month - the last ten days were all taken up by Rudder.

Today I am going to resume programming without a doubt. I want to find my center again and let loose once more. I had a good pace during the first 3 weeks of this month and now it is busted. I need to get it going.

To an extent I agree with the Deepmind guy who said that all that is needed is better memory. But it is actually necessary to go and do it, not just cram CPUs like the OpenAI guys had done.

Though Rudder wasn't exactly like I thought it would be, with this I've essentially covered the big 4: natural gradients, Zap, differentiable plasticity, reward distribution.

The tally would be 2.5/4. It is not too bad. I won't be doing RD, but I'll replace the MC for the actor with TD and do the updates using MC for the critic. If that works, I'll say that Rudder was a success.

Ahhhhh...there are no further gains to be dug up. It is all out in the open now and what will come from here on out is refinement.

It is the start of the game from here on out. I will go full out in mastering poker from here. It is only fitting that I start this out by a redesign of the tensor. It is a bit like starting from level one. It is important to not think of it as a chore.

I need to go at it as if I am starting a brand new game. As both DP and RD had unhinged me, I think I need to take some time to focus on more practical matters like the tensor redesign.

For the past five months I've been laboriously trying to gather the pieces before I start for real and now that the chasm is before me, I find myself scared in taking the plunge even though this is what I've been working towards all this time. It feel more like rather than learning new tricks, all I've really gotten is improved understanding of the basics.

Natural gradients are definitely a part of the basics. So is Hebbian metalearning. So is understanding reward distribution. A lot of what is cutting edge now will be undergraduate material in the future.

I do not yet understand hierarchical RL and unsupervised goal formation, but the pieces for that will land without a doubt. There will also be better ways of doing natural gradients than K-FAC and better ways of doing metalearning. Probably there won't be better ways of doing the contribution analysis - TD for the actor and MC or TD with large n will reign supreme.

Being able to make a strong poker player will be difficult without HRL with unsupervised goals - it in fact might be a prerequisite, but the present circumstances bind my hands.

The plan from here on out will be to get all the infrastructure and everything else in place. If it turns out to be not enough, then I'll have to do research in getting those pieces to click myself. But the community is making good progress on this, so I think I can just take my time to absolutely perfect the ML library. Maybe I will end up adding AD support in kernels just to simplify things - that will be an ideal opportunity to consider higher order AD as well. There is also so much that is missing like the convolutional operations. I'll also want to enable convolutional Hebbian learning and that means writing my own convolutional and deconvolutional kernels. There is a lot of engineering work waiting for me there that I am deliberately avoiding now.

Either way once HRL with unsupervised goals is done the prelude to the Singularity will be over and the Ascent to Transcendence project will start. As is usual, I won't strain myself over it. I got up way too early today just to read those pointless papers. I won't try to make up for perceived shortfalls by putting in even more effort.

The best way to get all the work done is to just keep a steady pace and keep it cool. I've been too much in the clouds lately.

It is time for a break. I've earned it.