Seriously, how long is this going to take me?
What I will do is split apply into pieces. One part calculates the offset changes and the other part filters out the pieces. I have no choice but to split it up. Also first I need to partition this into dimensions that are being processed and those that are not.

Let me take a break here. I'll deal with this once I am done. This is truly an amazing pain, but it won't let it hold me back.