I am putting quite a lot of thought into it as it turns out.
I have a fairly complete grasp on it now. I'll implement it tomorrow.

It is like the mental block that I had at the start of the day was just a mirage. It is quite amazing what application of proper programming principles can bring. Without a doubt, it will make a decisive factor in my victory.