I think I won't be doing programming today after all. In addition to NG, Zap, DP and Rudder, I think that GVFNs have enough potential that I need to study them in depth.
https://www.ijcai.org/Proceedings/13/Papers/246.pdf
Better Generalization with Forecasts

I've already read this paper.

https://sites.ualberta.ca/~amw8/phd.pdf
Developing a predictive approach to knowledge

Here is a 200 page PhD thesis on them.

After I am done reading this, I might consider asking Martha White for code samples on GVFNs. There is not enough information in the paper for me to quite get a feeling for them.

Let me get on with it.

These algorithmic innovation are absolutely essential in my view - they are irreplaceable and I cannot afford to overlook them. I cannot afford to wait for the community to converge on them either. I have to make my own way if I want to win this.