I am done with the thesis and the study of GVFNs. I find them quite interesting, but unfortunately for the time being I think I will leave them aside.
I do not think there is a good unsupervised way to learn GVFNs for RNN in the hidden layer right now so I will give this a pass. I'd be much better off focusin on Hebbian learning for the itme being.

I definitely think I have a feel for what these are doing now. It is definitely an interesting idea, but calling it a better memory might be a stretch.

Well, I do not get it right all the time.

As expected I now spent the most important part of my working day on something that did not pan out. It is just too bad. Well, let me do the chores for a bit and then I will do the apply for the ViewHostTensor.

I need to take a lesson from this. And the real lesson is that to not have to do research I have to know the truth - this will make falling into research traps hard to avoid.

These GVFNs will go the way of Bayesian reasoning when it comes to network design - they will be found to be hard to grasp from first principles.

What is needed is a grand theory of unsupervised learning - something that takes into consideration of how the network is occupied.

The noise in the brain is a cover for temperature (as in simulated annealing) and the sparsity is there to make credit assignment easier.

I need to grasp my focus again - my goal is only to win at poker for now. Get these RNNs done and then take advantage of the power of backpropagation to make that a reality.

I won't let myself be delayed anymore. This will be done by the end of the year.