At some point the time will come to dwell in the shadows and I will stop making these long commits, but until then I'll be as I am.
Right now there is great benefit in cooperating in the open and later there will be great benefit chasing individual goals in secret.

I've been pretty useless today all things considered. I managed to see through the GVFNs, but that had a cost attached to it. I am not the kind of guy that can fire spells non-stop. My mind seems to be operating under the vancian system.

I have some many worries and regrets - the only cure for that is constant steady effort as I have been doing up to now.

Tomorrow I will focus on mastering the view tensor. After I do that and finish the optimized RNN, I will make it my priority to test the TD actor idea.

Let me just focus on getting this done. After I am done with this batch of very overdue work, I will be able to breathe a sigh of relief.

Right now I am stuck in a hell of trying to reason through the optimization processes and that is not something I am capable of in the slightest. These things can only be tested and observed.

I need to focus on what I can do.

I did well for the day despite the outcome.

Today I broke the pace, but starting from here, I will resume where I left of yesterday and start cutting my attachments to research and to interacting on the social sites. I won't bother asking the authors any more questions and neither will I make those comments apart from making a summary in the PL monthly threads.

This is for the sake of getting the creative mania under control.

I'll focus on my core competency which is programming and work on extending that advantage there. There is no doubt that my capabilities are increasing from month to month.

I'll research plasticity, but that will be it. I won't bother chasing credit assignment schemes like Zap, Rudder and GVFNs. I'll stop worrying about that.

Let me just say one thing, higher order differentiation and indeed anything that requires multiple passes is pretty much dead. Once I add reversible learning that will add mutability during the forward pass. Since PRONG is already complicating things during the backward pass, that will pretty much kill that avenue of optimization.

Rudder which requires integrated gradients or relevance propagation was never a choice from the start and I knew it, but I was too scared to look the truth in the eye.

Differentiable programming is terror.

All those brilliant papers and that brilliant reasoning in them is just a kind of evolutionary search. There is no true insight to be found there - only heuristics.

Even if I ignored what is happening I doubt I would lose anything. Important discoveries will not get lost and the community will filter them to me. I must cast away the fear of falling behind - my only advantage is literally that I am doing this at my own pace and on my own time.

I've been doing this for over 3.5 years now ever since I started in early 2015. It is time to fulfill the ancient vows and bring the desires to light.

The only thing on my side is the language that I had created. The Spiral language is the symbol of my trust in the future.